refactor(provider/telegram): split events.ts into events/{inbound,reactions,index}
Mirrors the file layout used by other providers (iMessage, WhatsApp Business)
where each concern lives in its own sibling file. The previous events.ts had
grown to ~600 lines mixing three distinct concerns:

  - inbound.ts (437 LOC) — Telegram primitive → Spectrum content mappers
    (chatToSpace, userToSender, attachment/voice/contact/poll/richlink, the
    messageToContent dispatcher, and the top-level toTelegramMessage builder).
  - reactions.ts (98 LOC) — message_reaction → Spectrum reaction events,
    including the stub-target synthesis that wrapProviderMessage inflates.
  - index.ts (98 LOC) — public surface: MessagesOptions, the messages()
    stream factory, and the buildMessages() update dispatcher.

The public import (`import { messages } from "./events"`) is unchanged — Node
resolves it to events/index.ts. No behavior changes. ultracite + tsc clean
for the Telegram provider; pre-existing imessage/remote.ts errors remain
untouched.
